requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
seaborn
scikit-learn
matplotlib
# -e .